Add files via upload
This was very difficult, but only because I overanalyzed the situation far too much. The solution to making different branches of the survey was to systematically address conditions instead of trying to make the program check for all conditions at once. Maybe that's a more viable way to approach this as I get more skilled, thereby reducing the number of lines of code needed to achieve the same result, but this is how I felt it would be best to approach this for now.